requirements.txt
altgraph==0.17.2
asgiref==3.7.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.0
dj-database-url==2.2.0
django==4.2.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-jwt==1.19.2
future==0.18.2
gunicorn==23.0.0
idna==3.8
macholib==1.15.2
openai==0.28.0
packaging==24.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
six==1.15.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.2
Pillow==11.0.0